backend/requirements.txt
fastapi
uvicorn
langchain
pypdf
chromadb==0.4.24
python-dotenv
openai
ffmpeg-python
openai-whisper
python-multipart